Separa escucha del servidor de la app Elimina conexión innecesaria a la db
