+++ src/index.ts
@@ -1,24 +1,8 @@
-import express from 'express';
-import cors from 'cors';
-import routers from './apis/api';
-import 'reflect-metadata';
-import { createConnection } from 'typeorm';
+import app from "./app";
 
-const app = express();
 const port = 3000;
 
-createConnection();
-
-app.use(cors());
-app.use (express.json());
-app.use('/api', routers)
-
-app.get('/', (req, res) => {
-  res.send('Bienvenido')
-});
-
 app.listen(port, () => {
-  // console.log(`API REST Cakes Shop se escucha en http://localhost:${port}`);
+    console.log(`API REST Cakes Shop se escucha en http://localhost:${port}`);
 });
-
-export default app;
+
